requirements.txt
playwright>=1.40.0
asyncio-compat>=0.1.2
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0
schedule>=1.2.0
python-dotenv>=1.0.0
